notifications: split into energy-balance + body modules to fit cap
CI failure on the previous commit: server/lib/notifications.js grew to
823 lines after the noon + evening DB-backed report fixes, busting the
600-line hard cap that check:file-size --strict enforces. Split into
three focused modules:

- server/lib/energy-balance.js: pure bucketing math + DB fetch wrappers
  (computeOvernightStats, computeDailyStats, …). Mirrors what
  playground/js/energy-balance.js does for the in-app balance card.
- server/lib/notification-bodies.js: editorial body composers
  (buildEveningBody, buildNoonBody). Pure number → string.
- server/lib/notifications.js: alert/offline state, schedulers, and
  evaluate() — back under the cap at 595 lines.

Tests for the bucketing logic move out to tests/energy-balance.test.js
(plus added coverage for tankStoredEnergyKwh edge cases and
computeOvernightStats DB-error / db-absent paths). All 864 unit tests
pass; lint, knip, file-size --strict, check:assets all clean.
